Extract input_checked? to a module
