+++ actionpack/lib/action_view/helpers/tags/check_box.rb
@@ -1,7 +1,11 @@
+require 'action_view/helpers/tags/checkable'
+
 module ActionView
   module Helpers
     module Tags
       class CheckBox < Base #:nodoc:
+        include Checkable
+
         def initialize(object_name, method_name, template_object, checked_value, unchecked_value, options)
           @checked_value   = checked_value
           @unchecked_value = unchecked_value
